git-svn-id: http://47.100.187.235:13932/svn/ddxyz@235 9afdef4e-c05b-bb4b-bdb9-2143b111bd26

+++ src/main/java/cn/sowell/ddxyz/model/kanteen/dao/impl/KanteenWaresGroupDaoImpl.java
@@ -62,7 +62,7 @@ public class KanteenWaresGroupDaoImpl implements KanteenWaresGroupDao {
 		String sql = "update t_waresgroup_base set c_disabled = :toDisable where id = :waresgroupId";
 		SQLQuery query = sFactory.getCurrentSession().createSQLQuery(sql);
 		query.setLong("waresgroupId", waresGroupId);
-		query.setParameter("toDisable", toDisable?1:null, StandardBasicTypes.LONG);
+		query.setParameter("toDisable", toDisable?1:null, StandardBasicTypes.INTEGER);
 		return query.executeUpdate();
 	}
